// src/types/fishing/Fish.ts
export type Fish = {
    id: string;
    name: string;
    basic_price: number;
    bronze_price: number;
    silver_price: number;
    gold_price: number;
    osmium_price: number;
    location: string;
    time: string;
    weather: string;
    season: string;
    rarity: string;
    difficulty: string;
    offering: string;
    offering_quality: string;
};